Split apart some of the stuff in schema.js (#689)
